scripts/requirements_routeA.txt
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
contourpy==1.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
decorator==5.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
